Fix PHP notices

+++ applets/poll/twiml.php
@@ -2,15 +2,18 @@
 $ci =& get_instance();
 $poll = AppletInstance::getValue('poll');
 $option = AppletInstance::getValue('option');
-$number = normalize_phone_to_E164($_REQUEST['From']);
 
-$ci->db->delete('polls_responses', array('poll' => $poll, 'value' => $number));
-$ci->db->insert('polls_responses', array(
-	'poll' => $poll,
-	'value' => $number,
-	'response' => $option,
-	'time' => time()
-));
+if(!empty($_REQUEST['From'])) {
+	$number = normalize_phone_to_E164($_REQUEST['From']);
+
+	$ci->db->delete('polls_responses', array('poll' => $poll, 'value' => $number));
+	$ci->db->insert('polls_responses', array(
+		'poll' => $poll,
+		'value' => $number,
+		'response' => $option,
+		'time' => time()
+	));
+}
 
 $response = new Response();
